refactor: move live traffic related packages into traffic/livetraffic
