+++ integration/traffic/interface.go
@@ -1,6 +1,10 @@
 package traffic
 
-import "time"
+import (
+	"time"
+
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficproxy"
+)
 
 // HistoricalSpeedQuerier provide interfaces for Querying Historical Speeds
 type HistoricalSpeedQuerier interface {
@@ -9,4 +13,16 @@ type HistoricalSpeedQuerier interface {
 	// wayID: positive means travel forward, which is travel following edge's point sequence, negative means backward
 	// t: UTC time
 	QueryHistoricalSpeed(wayID int64, t time.Time) (float64, bool)
+}
+
+// LiveTrafficQuerier defines interfaces for querying traffic flows and incidents.
+type LiveTrafficQuerier interface {
+
+	// QueryFlow return the live traffic flow for a way
+	// wayID: positive means travel forward, which is travel following edge's point sequence, negative means backward
+	QueryFlow(wayID int64) *trafficproxy.Flow
+
+	// BlockedByIncident returns whether a way blocked by live traffic incident
+	// wayID: positive means travel forward, which is travel following edge's point sequence, negative means backward
+	BlockedByIncident(wayID int64) bool
 }
