+++ integration/traffic/livetraffic/trafficcache/flowscache/flows.go
@@ -3,7 +3,7 @@ package flowscache
 import (
 	"sync"
 
-	"github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficproxy"
 	"github.com/golang/glog"
 )
