+++ integration/traffic/livetraffic/trafficcache/flowscacheindexedbyedge/flows.go
@@ -6,7 +6,7 @@ import (
 	"github.com/Telenav/osrm-backend/integration/wayidsmap"
 
 	"github.com/Telenav/osrm-backend/integration/graph"
-	"github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficproxy"
 	"github.com/golang/glog"
 )
