+++ integration/traffic/livetraffic/trafficcache/flowscacheindexedbyedge/flows_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/Telenav/osrm-backend/integration/graph"
-	"github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficproxy"
 )
 
 func TestFlowsCache(t *testing.T) {
