+++ integration/traffic/livetraffic/trafficcache/incidentscache/incidents_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/Telenav/osrm-backend/integration/graph"
-	"github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficproxy"
 )
 
 func TestIncidentsCache(t *testing.T) {
