+++ integration/traffic/livetraffic/trafficcache/trafficcache/cache.go
@@ -1,9 +1,9 @@
 package trafficcache
 
 import (
-	"github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
-	"github.com/Telenav/osrm-backend/integration/trafficcache/flowscache"
-	"github.com/Telenav/osrm-backend/integration/trafficcache/incidentscache"
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficcache/flowscache"
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficcache/incidentscache"
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficproxy"
 	"github.com/golang/glog"
 )
