+++ integration/traffic/livetraffic/trafficdumper/dump.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
+	"github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficproxy"
 	"github.com/golang/glog"
 )
