+++ integration/traffic/livetraffic/trafficeater/interface.go
@@ -1,6 +1,6 @@
 package trafficeater
 
-import "github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
+import "github.com/Telenav/osrm-backend/integration/traffic/livetraffic/trafficproxy"
 
 // Eater is the interface that wraps the basic Eat method.
 type Eater interface {
